--- cmd/otto/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/rustyeddy/otto/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
