+++ cmd/ottoctl/cmd_cli_test.go
