+++ cmd/ottoctl/cmd_root_test.go
