+++ cmd/ottoctl/cmd_stats.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"log/slog"
 
-	"github.com/rustyeddy/otto/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -20,10 +19,7 @@ func statsRun(cmd *cobra.Command, args []string) {
 	// Check if we should connect to a remote server
 	client := GetClient()
 	if client == nil {
-		// Local mode: get stats directly
-		slog.Debug("Getting stats from local process")
-		stats := utils.GetStats()
-		fmt.Fprintf(cmdOutput, "Stats: %+v\n", stats)
+		fmt.Fprintf(cmdOutput, "Failed to get an otto client")
 		return
 	}
