+++ cmd/ottoctl/cmd_stats_test.go
