+++ cmd/ottoctl/cmd_version_test.go
