+++ cmd/ottoctl/ottoctl/main.go
@@ -1,0 +1,7 @@
+package main
+
+import cmd "github.com/rustyeddy/otto/cmd/ottoctl"
+
+func main() {
+	cmd.Execute()
+}
